organized mySql connector to its own file and routes
